+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Make example more usable
 - Implement node pinning with `controller.setPinned`
 - Implement node replacement with `controller.replaceNode`
+- Implement clearing the graph with `controller.clear`
 
 ## 0.1.1